coreos-install script fails with ubuntu 14.04
I had to hack the coreos-install script to get it to work from Ubuntu 14.04 Destop LiveCD.

For lsbk on line #208 I had to use the uppercase argument `P` instead of `p` and change from `partlabel` to `label`. Both those options were reported as unknown in the terminal. 

Because of the `partlabel` change, I also had to change from ${ROOT_DEV} to /dev/${ROOT_DEV} on line #211.

Not saying this is the correct way to resolve this issue (might break for other distros). But it should be clear where the errors were for me when executing the script. Hope it helps.